% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTypes.R
\name{copyTypes}
\alias{copyTypes}
\title{Copy Data Types}
\usage{
copyTypes(x, y)
}
\arguments{
\item{x}{"To" data frame}

\item{y}{"From" data frame}
}
\description{
Copy types from data frame y, to data frame x
}
\examples{
mtchar <- setTypes(mtcars,rep('character',11))
mtchar2 <- copyTypes(mtcars,mtchar)
}
\keyword{convenience}
\keyword{metaprogramming}
\keyword{types}
