% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeline.R
\name{getTimeline}
\alias{getTimeline}
\title{Makes Event Data Into Timeline}
\usage{
getTimeline(df, startVar, endVar, eventVar, start, end,
  groupingVar = NULL)
}
\arguments{
\item{df}{A data frame with columns date,dyad_name,dyad_new_id,deaths,and country.}

\item{start}{Start date}

\item{end}{End date}
}
\description{
Turns a dataset containing 1
}
\examples{
getTimeline(ged,'1999-01-01','2000-01-01') 
}
\keyword{dates}
\keyword{ged}
\keyword{timeline}
