% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randVect.R
\name{randVect}
\alias{randVect}
\title{Random Vector With Given Sum}
\usage{
randVect(N, M, sd = 1, pos.only = TRUE)
}
\arguments{
\item{N}{length of vector}

\item{M}{sum of vector}

\item{sd}{standard deviation}

\item{pos.only}{Only positive values?}
}
\description{
This function returns a vector of integers, with length N, and a sum of M
}
\examples{
v <- rand_vect(10,100) 
}
\keyword{random}
