% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTypes.R
\name{setTypes}
\alias{setTypes}
\title{Set Types By Vector}
\usage{
setTypes(df, types)
}
\arguments{
\item{df}{A data frame}

\item{types}{A character vector of length ncol(df)}
}
\description{
Set the data types of a data frame using a character vector.
The vector can be created by sapply(data,class)
}
\examples{
setTypes(mtcars,rep('character',11)) 
}
\keyword{metaprogramming}
\keyword{types}
